Move GFA extraction to gfa.cpp
