+++ gfa2gbwt.cpp
@@ -35,9 +35,9 @@ void parse_gfa(gbwt::GBWT& index, GBWTGraph& graph, const Config& config);
 void load_gbz(gbwt::GBWT& index, GBWTGraph& graph, const Config& config);
 void load_graph(gbwt::GBWT& index, GBWTGraph& graph, const Config& config);
 
-void write_gfa(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config);
-void write_gbz(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config);
-void write_graph(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config);
+void write_gfa(const GBWTGraph& graph, const Config& config);
+void write_gbz(const GBWTGraph& graph, const Config& config);
+void write_graph(const GBWTGraph& graph, const Config& config);
 
 void extract_translation(const GBWTGraph& graph, const Config& config);
 
@@ -79,15 +79,15 @@ main(int argc, char** argv)
   // Handle the output.
   if(config.output == output_gfa)
   {
-    write_gfa(index, graph, config);
+    write_gfa(graph, config);
   }
   else if(config.output == output_gbz)
   {
-    write_gbz(index, graph, config);
+    write_gbz(graph, config);
   }
   else if(config.output == output_graph)
   {
-    write_graph(index, graph, config);
+    write_graph(graph, config);
   }
 
   // Extract the translation.
@@ -308,10 +308,8 @@ load_graph(gbwt::GBWT& index, GBWTGraph& graph, const Config& config)
 
 //------------------------------------------------------------------------------
 
-// TODO this should be a library function with tests
-// TODO and progress information (number of each line type)
 void
-write_gfa(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config)
+write_gfa(const GBWTGraph& graph, const Config& config)
 {
   std::string gfa_name = config.basename + GFA_EXTENSION;
 
@@ -320,130 +318,12 @@ write_gfa(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config)
     std::cerr << "Writing the GFA to " << gfa_name << std::endl;
   }
   std::ofstream out(gfa_name, std::ios_base::binary);
-
-  // GFA header.
-  // TODO This should use buffered writing.
-  out << "H\tVN:Z:1.0\n";
-
-  // S-lines.
-  // TODO we should store views of all segment names for P-lines and W-lines
-  if(graph.has_segment_names())
-  {
-    graph.for_each_segment([&](const std::string& name, std::pair<nid_t, nid_t> nodes) -> bool
-    {
-      out << "S\t" << name << "\t";
-      for(nid_t id = nodes.first; id < nodes.second; id++)
-      {
-        out << graph.get_sequence(graph.get_handle(id, false));
-      }
-      out << "\n";
-      return true;
-    });
-  }
-  else
-  {
-    graph.for_each_handle([&](const handle_t& handle)
-    {
-      out << "S\t" << graph.get_id(handle) << "\t" << graph.get_sequence(handle) << "\n";
-    });
-  }
-
-  // L-lines.
-  if(graph.has_segment_names())
-  {
-    graph.for_each_link([&](const edge_t& edge, const std::string& from, const std::string& to) -> bool
-    {
-      out << "L\t"
-          << from << (graph.get_is_reverse(edge.first) ? "\t-\t" : "\t+\t")
-          << to << (graph.get_is_reverse(edge.second) ? "\t-\t" : "\t+\t")
-          << "*\n";
-      return true;
-    });
-  }
-  else
-  {
-    graph.for_each_edge([&](const edge_t& edge)
-    {
-      out << "L\t"
-          << graph.get_id(edge.first) << (graph.get_is_reverse(edge.first) ? "\t-\t" : "\t+\t")
-          << graph.get_id(edge.second) << (graph.get_is_reverse(edge.second) ? "\t-\t" : "\t+\t")
-          << "*\n";
-    });
-  }
-
-  // P-lines.
-  gbwt::size_type ref_sample = index.metadata.sample(REFERENCE_PATH_SAMPLE_NAME);
-  std::vector<gbwt::size_type> ref_paths = index.metadata.pathsForSample(ref_sample);
-  for(gbwt::size_type path_id : ref_paths)
-  {
-    gbwt::vector_type path = index.extract(gbwt::Path::encode(path_id, false));
-    out << "P\t" << index.metadata.contig(index.metadata.path(path_id).contig) << "\t";
-    size_t segments = path.size();
-    if(graph.has_segment_names())
-    {
-      // We assume that the path is a valid concatenation of segments.
-      size_t offset = 0;
-      while(offset < path.size())
-      {
-        auto segment = graph.get_segment(GBWTGraph::node_to_handle(path[offset]));
-        out << segment.first << (gbwt::Node::is_reverse(path[offset]) ? "-" : "+");
-        offset += segment.second.second - segment.second.first;
-        if(offset < path.size()) { out << ","; }
-      }
-    }
-    else
-    {
-      for(size_t i = 0; i < path.size(); i++)
-      {
-        out << gbwt::Node::id(path[i]) << (gbwt::Node::is_reverse(path[i]) ? "-" : "+");
-        if(i + 1 < path.size()) { out << ","; }
-      }
-    }
-    out << "\t";
-    for(size_t i = 1; i < segments; i++)
-    {
-      out << "*";
-      if(i + 1 < segments) { out << ","; }
-    }
-    out << "\n";
-  }
-
-  // W-lines.
-  for(gbwt::size_type path_id = 0; path_id < index.metadata.paths(); path_id++)
-  {
-    const gbwt::PathName& path_name = index.metadata.path(path_id);
-    if(path_name.sample == ref_sample) { continue; }
-    gbwt::vector_type path = index.extract(gbwt::Path::encode(path_id, false));
-    size_t length = 0;
-    for(auto node : path) { length += graph.get_length(GBWTGraph::node_to_handle(node)); }
-    out << "W\t" << index.metadata.sample(path_name.sample)
-        << "\t" << path_name.phase
-        << "\t" << index.metadata.contig(path_name.contig)
-        << "\t" << path_name.count << "\t" << (path_name.count + length) << "\t";
-    if(graph.has_segment_names())
-    {
-      // We assume that the path is a valid concatenation of segments.
-      size_t offset = 0;
-      while(offset < path.size())
-      {
-        auto segment = graph.get_segment(GBWTGraph::node_to_handle(path[offset]));
-        out << (gbwt::Node::is_reverse(path[offset]) ? "<" : ">") << segment.first;
-        offset += segment.second.second - segment.second.first;
-      }
-    }
-    else
-    {
-      for(auto node : path)
-      {
-        out << (gbwt::Node::is_reverse(node) ? "<" : ">") << gbwt::Node::id(node);
-      }
-    }
-    out << "\n";
-  }
+  gbwt_to_gfa(graph, out, config.show_progress);
+  out.close();
 }
 
 void
-write_gbz(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config)
+write_gbz(const GBWTGraph& graph, const Config& config)
 {
   std::string gbz_name = config.basename + GBWTGraph::COMPRESSED_EXTENSION;
 
@@ -457,13 +337,13 @@ write_gbz(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config)
     std::cerr << "gfa2gbwt: Cannot open file " << gbz_name << " for writing" << std::endl;
     std::exit(EXIT_FAILURE);
   }
-  index.serialize(out);
+  graph.index->serialize(out);
   graph.compress(out);
   out.close();
 }
 
 void
-write_graph(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& config)
+write_graph(const GBWTGraph& graph, const Config& config)
 {
   std::string gbwt_name = config.basename + gbwt::GBWT::EXTENSION;
   std::string graph_name = config.basename + GBWTGraph::EXTENSION;
@@ -472,7 +352,7 @@ write_graph(const gbwt::GBWT& index, const GBWTGraph& graph, const Config& confi
   {
     std::cerr << "Writing GBWT to " << gbwt_name << std::endl;
   }
-  if(!sdsl::store_to_file(index, gbwt_name))
+  if(!sdsl::store_to_file(*(graph.index), gbwt_name))
   {
     std::cerr << "gfa2gbwt: Cannot write GBWT to " << gbwt_name << std::endl;
     std::exit(EXIT_FAILURE);
